project7-TestPlan_AI_AGENT_JIRA/intelligent-test-plan-agent/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
aiohttp==3.11.0
pypdf==5.1.0
groq==0.11.0
python-dotenv==1.0.0
keyring==25.5.0
keyrings.alt==5.0.0
httpx==0.27.0
